requirements.txt
amqp==5.3.1
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.1
Brotli==1.1.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.3
cssselect2==0.7.0
defusedxml==0.8.0rc2
dj-database-url==2.3.0
Django==5.1.2
django-celery-beat==2.7.0
django-environ==0.11.2
django-oauth-toolkit==3.0.1
django-timezone-field==7.0
django-widget-tweaks==1.5.0
fonttools==4.54.1
google-ai-generativelanguage==0.6.15
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.4
googleapis-common-protos==1.65.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
jwcrypto==1.5.6
kombu==5.4.2
lxml==5.3.0
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.10
psycopg2cffi==2.9.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pyparsing==3.2.0
pypdf==5.1.0
pyphen==0.15.0
python-bidi==0.6.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-social-auth==0.3.6
python3-openid==3.2.0
PyYAML==6.0.2
qrcode==8.0
RapidFuzz==3.10.1
razorpay==1.4.2
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.0
webencodings==0.5.1
whitenoise==6.8.2
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
tensorflow>=2.5.0
Pillow>=8.0.0
numpy>=1.19.0
